refactor(advisor): extract builders module, data-drive freetext flag
Move buildProgress/Caveats/Why/Nav/VerdictHTML and the icon
infrastructure (ICON_SVGS, initIcons, ic) to advisor-builders.js;
widget imports them. buildNav now takes state as an explicit arg.
buildProgress receives questions directly instead of closing over TOTAL.

Replace the hardcoded qId==='backend'&&value==='other' freetext check
with a data-driven selectedOpt?.freetext flag, mirroring the existing
escape pattern. Adds freetext:true to the YAML option.
